Add files via upload
How to Use:
Need to import glob for opening images in folder.
To convert JPG's to GIF we'll use Pillow so we are using from PIL import Image command from pillow library.
Be aware that your images are all same size and all are JPG's.
create a folder with jpg images which you want to convert to GIF.
don't forget to put your folder adress to make_gif("your/jpg/folder") command.
then run the main.py
When you execute the main.py from the command prompt you need to put image duration. For example you can put 50. If you want to make GIF slower, increase the value of duration.
That is all.